implemented atomicbusmanager
wrote skeleton functions and comments for the other functions needed /
make skeleton classes that need to be fleshed out to actually have
everything work.

+++ 418Cache/418Cache/CacheConstants.cpp
@@ -56,7 +56,8 @@ int CacheConstants::getNumBytesInLine(){
 int CacheConstants::getNumCacheSize(){
 	return numCacheSize;
 }
-int CacheConstants::getCycle(){
+
+unsigned long long CacheConstants::getCycle(){
 	return cycles;
 }
